+++ yf-project-api/src/main/java/org/opensource/model/response/financials/CostOfRevenue.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class CostOfRevenue{
 
 	@JsonProperty("raw")
-	private Integer raw;
+	private Long raw;
 
 	@JsonProperty("longFmt")
 	private String longFmt;
@@ -13,7 +13,7 @@ public class CostOfRevenue{
 	@JsonProperty("fmt")
 	private String fmt;
 
-	public Integer getRaw(){
+	public Long getRaw(){
 		return raw;
 	}
