+++ yf-project-api/src/main/java/org/opensource/model/response/keystats/BidSize.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class BidSize{
 
 	@JsonProperty("raw")
-	private Integer raw;
+	private Long raw;
 
 	@JsonProperty("longFmt")
 	private String longFmt;
@@ -13,7 +13,7 @@ public class BidSize{
 	@JsonProperty("fmt")
 	private String fmt;
 
-	public Integer getRaw(){
+	public Long getRaw(){
 		return raw;
 	}
